requirements.txt
fastapi>=0.119.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
pydantic-settings>=2.8.1,<3.0.0
python-multipart>=0.0.9,<1.0.0
transformers>=4.51.0,<5.0.0
huggingface-hub>=0.23.0,<1.0.0
urllib3<=2.5.0
torch>=2.5.0,<3.0.0
sentencepiece>=0.2.0,<1.0.0
pypdf>=5.3.0,<6.0.0
python-docx>=1.1.2,<2.0.0
langdetect>=1.0.9,<2.0.0
numpy>=1.26.0,<2.0.0
scipy>=1.11.0,<1.13.0
scikit-learn>=1.4.0,<1.6.0
